Unkstruct_801C3F9C

+++ src/st/nz0/3E30C.c
@@ -1106,9 +1106,9 @@ void func_801C3F9C(Unkstruct_801C3F9C** self) {
 
     case 1:
         temp = (*self)->unk0A + ((*self)->unk1E / 3);
-        g_api.CheckCollision((*self)->unk14, (s16)(temp), &collider, 0);
+        g_api.CheckCollision((*self)->unk14, temp, &collider, 0);
         if (collider.unk0 % 2) {
-            (*self)->unk0A = (u16)(*self)->unk0A + collider.unk18;
+            (*self)->unk0A = (*self)->unk0A + collider.unk18;
             if ((*self)->unk10 < 0x4000) {
                 (*self)->unk2C = 1;
             }
